+++ index.html
@@ -164,7 +164,7 @@
       </div>
     </section>
   </template>
-<script scr="js/main.js">
+<script scr="js/main.js" type="module">
 </script>
 </body>
 </html>
